+++ tencentcloud/batch/v20170312/models.go
@@ -80,16 +80,20 @@ type AnonymousComputeEnv struct {
 }
 
 type Application struct {
-	// 任务执行命令
-	Command *string `json:"Command,omitempty" name:"Command"`
-
 	// 应用程序的交付方式，包括PACKAGE、LOCAL 两种取值，分别指远程存储的软件包、计算环境本地。
+	// 注意：此字段可能返回 null，表示取不到有效值。
 	DeliveryForm *string `json:"DeliveryForm,omitempty" name:"DeliveryForm"`
 
+	// 任务执行命令。与Commands不能同时指定。
+	// 注意：此字段可能返回 null，表示取不到有效值。
+	Command *string `json:"Command,omitempty" name:"Command"`
+
 	// 应用程序软件包的远程存储路径
+	// 注意：此字段可能返回 null，表示取不到有效值。
 	PackagePath *string `json:"PackagePath,omitempty" name:"PackagePath"`
 
 	// 应用使用Docker的相关配置。在使用Docker配置的情况下，DeliveryForm 为 LOCAL 表示直接使用Docker镜像内部的应用软件，通过Docker方式运行；DeliveryForm 为 PACKAGE，表示将远程应用包注入到Docker镜像后，通过Docker方式运行。为避免Docker不同版本的兼容性问题，Docker安装包及相关依赖由Batch统一负责，对于已安装Docker的自定义镜像，请卸载后再使用Docker特性。
+	// 注意：此字段可能返回 null，表示取不到有效值。
 	Docker *Docker `json:"Docker,omitempty" name:"Docker"`
 }