+++ tencentcloud/cdb/v20170320/models.go
@@ -9356,7 +9356,7 @@ type InstanceAuditLogFilters struct {
 	// errCode - 错误码；
 	// threadId - 线程ID；
 	// 
-	// 范围搜索（时间类型统一为微妙）：
+	// 范围搜索（时间类型统一为微秒）：
 	// execTime - 执行时间；
 	// lockWaitTime - 执行时间；
 	// ioWaitTime - IO等待时间；