+++ tencentcloud/redis/v20180412/models.go
@@ -3580,10 +3580,17 @@ type DescribeInstancesRequestParams struct {
 	// 
 	InstanceId *string `json:"InstanceId,omitempty" name:"InstanceId"`
 
-	// 实例列表排序依据，枚举值如下所示：<ul><li>projectId：依据项目ID排序。</li><li>createtime：依据实例创建时间排序。</li><li>instancename：依据实例名称排序。</li><li>type：依据实例类型排序。</li><li>curDeadline：依据实例到期时间排序。</li></ul>
+	// 实例列表排序依据，枚举值如下所示：
+	// - projectId：依据项目ID排序。
+	// - createtime：依据实例创建时间排序。
+	// - instancename：依据实例名称排序。
+	// - type：依据实例类型排序。
+	// - curDeadline：依据实例到期时间排序。
 	OrderBy *string `json:"OrderBy,omitempty" name:"OrderBy"`
 
-	// 实例排序方式，默认为倒序排序。<ul><li>1：倒序。</li><li>0：顺序。</li></ul>
+	// 实例排序方式，默认为倒序排序。
+	// - 1：倒序。
+	// - 0：顺序。
 	OrderType *int64 `json:"OrderType,omitempty" name:"OrderType"`
 
 	// 私有网络 ID 数组。如果不配置该参数或设置数组为空则默认选择基础网络。例如47525。该参数暂时保留，可忽略。请根据 UniqVpcIds 参数格式设置私有网络ID数组。
@@ -3592,7 +3599,7 @@ type DescribeInstancesRequestParams struct {
 	// 私有网络所属子网 ID 数组，例如：56854。该参数暂时保留，可忽略。请根据 UniqSubnetIds 参数格式设置私有网络子网 ID 数组。
 	SubnetIds []*string `json:"SubnetIds,omitempty" name:"SubnetIds"`
 
-	// 设置模糊查询关键字，支持根据实例名称或实例ID模糊查询实例。
+	// 设置模糊查询关键字段，仅实例名称支持模糊查询。
 	SearchKey *string `json:"SearchKey,omitempty" name:"SearchKey"`
 
 	// 项目 ID 组成的数组。
@@ -3610,19 +3617,32 @@ type DescribeInstancesRequestParams struct {
 	// 地域 ID 数组，该参数已经弃用，可通过公共参数Region查询对应地域。
 	RegionIds []*int64 `json:"RegionIds,omitempty" name:"RegionIds"`
 
-	// 实例状态。<ul><li>0：待初始化。</li><li>1：流程中。</li><li>2：运行中。</li><li>-2：已隔离。</li><li>-3：待删除。</li></ul>
+	// 实例状态。
+	// - 0：待初始化。
+	// - 1：流程中。
+	// - 2：运行中。
+	// - -2：已隔离。
+	// - -3：待删除。
 	Status []*int64 `json:"Status,omitempty" name:"Status"`
 
-	// 实例架构版本。<ul><li>1：单机版。</li><li>2：主从版。</li><li>3：集群版。</li></ul>
+	// 实例架构版本。
+	// - 1：单机版。
+	// - 2：主从版。
+	// - 3：集群版。
 	TypeVersion *int64 `json:"TypeVersion,omitempty" name:"TypeVersion"`
 
 	// 存储引擎信息。可设置为Redis-2.8、Redis-4.0、Redis-5.0、Redis-6.0 或者 CKV。
 	EngineName *string `json:"EngineName,omitempty" name:"EngineName"`
 
-	// 续费模式。<ul><li>0：手动续费。</li><li>1：自动续费。</li><li>2：到期不再续费。</ul>
+	// 续费模式。
+	// - 0：手动续费。
+	// - 1：自动续费。
+	// - 2：到期不再续费。
 	AutoRenew []*int64 `json:"AutoRenew,omitempty" name:"AutoRenew"`
 
-	// 计费模式。<ul><li>postpaid：按量计费。</li><li>prepaid：包年包月。</li></ul>
+	// 计费模式。
+	// - postpaid：按量计费。
+	// - prepaid：包年包月。
 	BillingMode *string `json:"BillingMode,omitempty" name:"BillingMode"`
 
 	// 实例类型。
@@ -3638,7 +3658,7 @@ type DescribeInstancesRequestParams struct {
 	// - 16：Redis 6.2内存版（集群架构）。
 	Type *int64 `json:"Type,omitempty" name:"Type"`
 
-	// 设置搜索关键字数组，可根据实例ID、实例名称、完整IP地址查询实例。
+	// 设置搜索关键字数组，可根据实例ID、实例名称、完整IP地址搜索实例。
 	SearchKeys []*string `json:"SearchKeys,omitempty" name:"SearchKeys"`
 
 	// 内部参数，用户可忽略。
@@ -3653,13 +3673,17 @@ type DescribeInstancesRequestParams struct {
 	// 根据标签的 Key 筛选资源，该参数不配置或者数组设置为空值，则不根据标签Key进行过滤。
 	TagKeys []*string `json:"TagKeys,omitempty" name:"TagKeys"`
 
-	// 实例的产品版本。如果该参数不配置或者数组设置为空值，则默认不依据此参数过滤实例。<ul><li>local：本地盘版。</li><li>cdc：独享集群版。</li></ul>
+	// 实例的产品版本。如果该参数不配置或者数组设置为空值，则默认不依据此参数过滤实例。
+	// - local：本地盘版。
+	// - cdc：独享集群版。
 	ProductVersions []*string `json:"ProductVersions,omitempty" name:"ProductVersions"`
 
 	// 批量查询指定的实例 ID，返回结果已 Limit 限制为主。
 	InstanceIds []*string `json:"InstanceIds,omitempty" name:"InstanceIds"`
 
-	// 可用区模式。<ul><li>singleaz：单可用区。</li><li>multiaz：多可用区。</li></ul>
+	// 可用区模式。
+	// - singleaz：单可用区。
+	// - multiaz：多可用区。
 	AzMode *string `json:"AzMode,omitempty" name:"AzMode"`
 }
 
@@ -3676,10 +3700,17 @@ type DescribeInstancesRequest struct {
 	// 
 	InstanceId *string `json:"InstanceId,omitempty" name:"InstanceId"`
 
-	// 实例列表排序依据，枚举值如下所示：<ul><li>projectId：依据项目ID排序。</li><li>createtime：依据实例创建时间排序。</li><li>instancename：依据实例名称排序。</li><li>type：依据实例类型排序。</li><li>curDeadline：依据实例到期时间排序。</li></ul>
+	// 实例列表排序依据，枚举值如下所示：
+	// - projectId：依据项目ID排序。
+	// - createtime：依据实例创建时间排序。
+	// - instancename：依据实例名称排序。
+	// - type：依据实例类型排序。
+	// - curDeadline：依据实例到期时间排序。
 	OrderBy *string `json:"OrderBy,omitempty" name:"OrderBy"`
 
-	// 实例排序方式，默认为倒序排序。<ul><li>1：倒序。</li><li>0：顺序。</li></ul>
+	// 实例排序方式，默认为倒序排序。
+	// - 1：倒序。
+	// - 0：顺序。
 	OrderType *int64 `json:"OrderType,omitempty" name:"OrderType"`
 
 	// 私有网络 ID 数组。如果不配置该参数或设置数组为空则默认选择基础网络。例如47525。该参数暂时保留，可忽略。请根据 UniqVpcIds 参数格式设置私有网络ID数组。
@@ -3688,7 +3719,7 @@ type DescribeInstancesRequest struct {
 	// 私有网络所属子网 ID 数组，例如：56854。该参数暂时保留，可忽略。请根据 UniqSubnetIds 参数格式设置私有网络子网 ID 数组。
 	SubnetIds []*string `json:"SubnetIds,omitempty" name:"SubnetIds"`
 
-	// 设置模糊查询关键字，支持根据实例名称或实例ID模糊查询实例。
+	// 设置模糊查询关键字段，仅实例名称支持模糊查询。
 	SearchKey *string `json:"SearchKey,omitempty" name:"SearchKey"`
 
 	// 项目 ID 组成的数组。
@@ -3706,19 +3737,32 @@ type DescribeInstancesRequest struct {
 	// 地域 ID 数组，该参数已经弃用，可通过公共参数Region查询对应地域。
 	RegionIds []*int64 `json:"RegionIds,omitempty" name:"RegionIds"`
 
-	// 实例状态。<ul><li>0：待初始化。</li><li>1：流程中。</li><li>2：运行中。</li><li>-2：已隔离。</li><li>-3：待删除。</li></ul>
+	// 实例状态。
+	// - 0：待初始化。
+	// - 1：流程中。
+	// - 2：运行中。
+	// - -2：已隔离。
+	// - -3：待删除。
 	Status []*int64 `json:"Status,omitempty" name:"Status"`
 
-	// 实例架构版本。<ul><li>1：单机版。</li><li>2：主从版。</li><li>3：集群版。</li></ul>
+	// 实例架构版本。
+	// - 1：单机版。
+	// - 2：主从版。
+	// - 3：集群版。
 	TypeVersion *int64 `json:"TypeVersion,omitempty" name:"TypeVersion"`
 
 	// 存储引擎信息。可设置为Redis-2.8、Redis-4.0、Redis-5.0、Redis-6.0 或者 CKV。
 	EngineName *string `json:"EngineName,omitempty" name:"EngineName"`
 
-	// 续费模式。<ul><li>0：手动续费。</li><li>1：自动续费。</li><li>2：到期不再续费。</ul>
+	// 续费模式。
+	// - 0：手动续费。
+	// - 1：自动续费。
+	// - 2：到期不再续费。
 	AutoRenew []*int64 `json:"AutoRenew,omitempty" name:"AutoRenew"`
 
-	// 计费模式。<ul><li>postpaid：按量计费。</li><li>prepaid：包年包月。</li></ul>
+	// 计费模式。
+	// - postpaid：按量计费。
+	// - prepaid：包年包月。
 	BillingMode *string `json:"BillingMode,omitempty" name:"BillingMode"`
 
 	// 实例类型。
@@ -3734,7 +3778,7 @@ type DescribeInstancesRequest struct {
 	// - 16：Redis 6.2内存版（集群架构）。
 	Type *int64 `json:"Type,omitempty" name:"Type"`
 
-	// 设置搜索关键字数组，可根据实例ID、实例名称、完整IP地址查询实例。
+	// 设置搜索关键字数组，可根据实例ID、实例名称、完整IP地址搜索实例。
 	SearchKeys []*string `json:"SearchKeys,omitempty" name:"SearchKeys"`
 
 	// 内部参数，用户可忽略。
@@ -3749,13 +3793,17 @@ type DescribeInstancesRequest struct {
 	// 根据标签的 Key 筛选资源，该参数不配置或者数组设置为空值，则不根据标签Key进行过滤。
 	TagKeys []*string `json:"TagKeys,omitempty" name:"TagKeys"`
 
-	// 实例的产品版本。如果该参数不配置或者数组设置为空值，则默认不依据此参数过滤实例。<ul><li>local：本地盘版。</li><li>cdc：独享集群版。</li></ul>
+	// 实例的产品版本。如果该参数不配置或者数组设置为空值，则默认不依据此参数过滤实例。
+	// - local：本地盘版。
+	// - cdc：独享集群版。
 	ProductVersions []*string `json:"ProductVersions,omitempty" name:"ProductVersions"`
 
 	// 批量查询指定的实例 ID，返回结果已 Limit 限制为主。
 	InstanceIds []*string `json:"InstanceIds,omitempty" name:"InstanceIds"`
 
-	// 可用区模式。<ul><li>singleaz：单可用区。</li><li>multiaz：多可用区。</li></ul>
+	// 可用区模式。
+	// - singleaz：单可用区。
+	// - multiaz：多可用区。
 	AzMode *string `json:"AzMode,omitempty" name:"AzMode"`
 }